Replacement Pages
I started to work on the replacement pages that will be used to replace the main.html page and the main.css page. So far I have begun little work, with the idea of making a more improved version of the website so that I can use it as a way to practice what I learned.